requirements.txt
astroid==2.3.3
autopep8==1.4.4
colorama==0.4.1
coverage==4.5.4
dj-database-url==0.5.0
Django==3.0.3
django-js-asset==1.2.2
et-xmlfile==1.0.1
gunicorn==20.0.0
idna==2.8
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
lxml==4.4.1
mccabe==0.6.1
mock==3.0.5
multidict==4.5.2
pycodestyle
pylint==2.4.3
pylint-django==2.0.12
pylint-plugin-utils==0.6
python-docx==0.8.10
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
whitenoise==4.1.4
wrapt==1.11.2
yarl==1.2.6
psycopg2
Pillow
django-db-file-storage
django-rest-framework
django-rest-knox
django-cors-headers